mobile: the rendezvous captures turn on their side
Five dashboards stacked one above the next at 390 was 1,361px of scrolling
to look at an argument that is one glance wide. Below 620 the set is a
horizontal row of 78vw cards with snap points, native overflow and the
deviation card's edge fades, promoted here into a pattern anything else
with a set of frames can use.

Two elements of markup and both are display:contents above the breakpoint,
so the desktop layout is not merely close but the same boxes: a full-page
1440 capture diffs against the pre-batch one at zero pixels.

The contrast walk moved out of verify into scripts/lib/audit.mjs so the
narrow widths can be cleared per unit rather than per pass.

390: rendezvous 1,577 -> 675, playground 4,144 -> 3,242. Saved 902px.
